+++ CHANGES.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add debug option to show strength hint for all player territories
 * Add program flag `--ai-debug` to print what the AI is thinking
 * Add test islands to persist test maps in VCS
+* Add shortcut `ctrl` + `N` to create a new island in the map editor
 
 ### Changed